requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
music21==9.1.0
soundsliceapi
python-dotenv==1.0.0
gunicorn==21.2.0
pymongo==4.6.1
werkzeug==3.0.1
numpy==1.26.2